requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.0
imageio==2.33.0
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pygame==2.5.2
pygifsicle==1.0.7
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
support-developer==1.0.5
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.1
opencv-python==4.8.1.78
Pillow==10.0.1
proglog==0.1.10
pygame==2.5.2
requests==2.31.0
setuptools==68.2.2
tqdm==4.66.1
urllib3==2.0.7
